+++ transform/models/refined_zone/stats_publiques/stats_annuelles/annual_quantity_processed_by_bordereaux_type.sql
@@ -1,7 +1,7 @@
 select
     extract(
         'year' from date_trunc('year', processed_at)
-    ) as annee,
+    )::int as annee,
     _bs_type as type_bordereau,
     sum(
         case
